+++ handshape_datasets/datasets/nus1.py
@@ -130,5 +130,8 @@ class Nus1(DatasetLoader):
         if (len(npz_exist) == 0):
             return False
         else:
-            rmtree(folder)
+            if (os.path.exists(folder)):
+                rmtree(folder)
+            else:
+                return False
             return True
