+++ handshape_datasets/datasets/rwth.py
@@ -98,5 +98,8 @@ class RWTH(DatasetLoader):
         if (len(npz_exist) == 0):
             return False
         else:
-            rmtree(folder)
+            if (os.path.exists(folder)):
+                rmtree(folder)
+            else:
+                return False
             return True
